separate training and prediction tests
